+++ lib/Model/DlibCnnModel/DlibCnn5Model.php
@@ -1,8 +1,9 @@
 <?php
 /**
  * @copyright Copyright (c) 2020, Matias De lellis <mati86dl@gmail.com>
+ * @copyright Copyright (c) 2018, Branko Kokanovic <branko@kokanovic.org>
  *
- * @author Matias De lellis <mati86dl@gmail.com>
+ * @author Branko Kokanovic <branko@kokanovic.org>
  *
  * @license GNU AGPL version 3 or any later version
  *
@@ -21,31 +22,34 @@
  *
  */
 
-namespace OCA\FaceRecognition\Model;
+namespace OCA\FaceRecognition\Model\DlibCnnModel;
 
-class DlibCnn68Model extends DlibCnnModel implements IModel {
+use OCA\FaceRecognition\Model\DlibCnnModel\DlibCnnModel;
+use OCA\FaceRecognition\Model\IModel;
+
+class DlibCnn5Model extends DlibCnnModel implements IModel {
 
 	/** Defines ID for default face model */
-	const FACE_MODEL_ID = 2;
+	const FACE_MODEL_ID = 1;
 
 	/** Defines name for default face model */
-	const FACE_MODEL_NAME = 'DlibCnn68';
+	const FACE_MODEL_NAME = 'Default';
 
 	/** Defines description for default face model */
-	const FACE_MODEL_DESC = 'Alternative default model, using dlib: mmod_human_face_detector.dat, shape_predictor_68_face_landmarks.dat and dlib_face_recognition_resnet_model_v1.dat';
+	const FACE_MODEL_DESC = 'Main model, using dlib defaults: mmod_human_face_detector.dat, shape_predictor_5_face_landmarks.dat and dlib_face_recognition_resnet_model_v1.dat';
 
 	/*
 	 * Model files.
 	 */
 	const FACE_MODEL_BZ2_URLS = [
 		'https://github.com/davisking/dlib-models/raw/94cdb1e40b1c29c0bfcaf7355614bfe6da19460e/mmod_human_face_detector.dat.bz2',
-		'https://github.com/davisking/dlib-models/raw/4af9b776281dd7d6e2e30d4a2d40458b1e254e40/shape_predictor_68_face_landmarks.dat.bz2',
+		'https://github.com/davisking/dlib-models/raw/4af9b776281dd7d6e2e30d4a2d40458b1e254e40/shape_predictor_5_face_landmarks.dat.bz2',
 		'https://github.com/davisking/dlib-models/raw/2a61575dd45d818271c085ff8cd747613a48f20d/dlib_face_recognition_resnet_model_v1.dat.bz2'
 	];
 
 	const FACE_MODEL_FILES = [
 		'mmod_human_face_detector.dat',
-		'shape_predictor_68_face_landmarks.dat',
+		'shape_predictor_5_face_landmarks.dat',
 		'dlib_face_recognition_resnet_model_v1.dat'
 	];
